requirements.txt
requests==2.31.0
google-api-python-client==2.93.0
google-auth==2.23.0
python-dateutil==2.8.2
